services

+++ src/main/java/com/sec/services/communication/MessageCrudInterface.java
@@ -7,5 +7,5 @@ import com.sec.model.comunication.Message;
 public interface MessageCrudInterface {
 Message addMessage(Message m);
 List<Message> showListMessage();
-void DeleteMessage(int id);
+void DeleteMessage(long id);
 }
